+++ craveThru/craveThru/PlacesAPI/File.swift
@@ -1,0 +1,9 @@
+//
+//  File.swift
+//  craveThru
+//
+//  Created by Raymond Esteybar on 5/6/19.
+//  Copyright © 2019 Eros Gonzalez. All rights reserved.
+//
+
+import Foundation
